RPC cache: Use serail number approach for detecting if cache item is not fresh
# HG changeset patch
# User Dan Kruchinin <dkruchinin@acm.org>
# Date 1314375277 -10800
# Node ID f8424e827b4c5f3c8322fc20cc773cd4cf0b1447
# Parent  0077b14fc71f8513cd99fae35cff8c05d9456184
RPC cache: Use serail number approach for detecting if cache item is not fresh.

RPC binding has a serial number (16 bit integer) that is incremented
every time binding is updated. Each nlm_rpc_t in cache has a copy of the
serial number of RPC binding at the moment it is initialized. In order
to detect if cache item is not "fresh" (i.e. it was initialized with old
binding), we just compare its serial number with a serial number of
a new binding.

Signed-off-by: Dan Kruchinin <dkruchinin@acm.org>